backend/requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
audioread==3.1.0
bcrypt==5.0.0
boto3==1.38.46
botocore==1.38.46
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
decorator==5.2.1
ecdsa==0.19.1
exceptiongroup==1.3.1
fastapi==0.124.4
filelock==3.20.0
fsspec==2025.12.0
greenlet==3.3.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.13.4
HyperPyYAML==1.2.2
idna==3.11
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.5.3
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
loguru==0.7.3
Mako==1.3.10
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.1.2
networkx==3.4.2
numba==0.63.1
numpy==1.23.5
packaging==25.0
passlib==1.7.4
platformdirs==4.5.1
pooch==1.8.2
pyasn1==0.4.8
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.4.0
python-multipart==0.0.21
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
s3transfer==0.13.1
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
speechbrain==1.0.3
SQLAlchemy==2.0.45
psycopg2-binary==2.9.9
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
tomli==2.3.0
torch==2.1.2
torchaudio==2.1.2
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.38.0
win32_setctime==1.2.0


